% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_major_incidents.R
\name{flag_major_incidents}
\alias{flag_major_incidents}
\title{Flag or Filter Major Wildfire Incidents}
\usage{
flag_major_incidents(data, threshold = 50, filter = TRUE)
}
\arguments{
\item{data}{A data frame with `estimated_financial_loss_million` column.}

\item{threshold}{Minimum financial loss to qualify as major (default = 50).}

\item{filter}{If TRUE, return only major incidents. If FALSE, return full data with `is_major` column.}
}
\value{
A data frame.
}
\description{
Flags or filters major incidents based on financial loss.
}
