% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data.R
\name{inspect_data}
\alias{inspect_data}
\title{Inspect Structure and Missing Values}
\usage{
inspect_data(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A list with structure and missingness summary.
}
\description{
Inspects the structure and shows missing value counts for each column.
}
